fix: resolve GitHub issues #58, #80, #108, #133, #167
#167 — Splash screen: replace isDark ternary with theme.colorScheme.surface
  to fix hardcoded color that broke dark mode.

#80 — Chat: remove kCurrentUserIdStub; wire real currentUserProvider from
  Supabase auth into send-controller and chat-thread-screen; extract
  duplicate string constants in chat_thread_screen.

#108 — Icons: replace all raw size: N literals with DeelmarktIconSize tokens
  across 22 widget files; replace SizedBox magic numbers with Spacing tokens;
  refactor long build() methods in conversation_list_empty_state and
  search_initial_view to stay under the 60-line limit.

#133 — Test files: split oversized test files (1000+ lines) into focused
  groups (auth repository split into 5 files, pr33 split into 4 files);
  fix TestWidgetsFlutterBinding init in language-switch test.

#58 — Search: wire toggleFavourite in SearchNotifier with optimistic UI;
  move toggleFavouriteUseCaseProvider to core/services/repository_providers
  to fix CROSS_IMPORT violation; add missing widget stub tests.
