+++ lib/features/admin/presentation/widgets/admin_narrow_viewport_message.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 
 import 'package:deelmarkt/core/design_system/colors.dart';
+import 'package:deelmarkt/core/design_system/icon_sizes.dart';
 import 'package:deelmarkt/core/design_system/spacing.dart';
 
 /// Shown in place of the admin shell when the viewport is too narrow
@@ -25,7 +26,7 @@ class AdminNarrowViewportMessage extends StatelessWidget {
               children: [
                 const Icon(
                   PhosphorIconsRegular.desktopTower,
-                  size: 48,
+                  size: DeelmarktIconSize.xl,
                   color: DeelmarktColors.neutral300,
                 ),
                 const SizedBox(height: Spacing.s4),
