+++ lib/features/auth/presentation/widgets/kyc_banner.dart
@@ -2,6 +2,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 
+import 'package:deelmarkt/core/design_system/icon_sizes.dart';
 import 'package:deelmarkt/core/design_system/spacing.dart';
 import 'package:deelmarkt/core/design_system/trust_theme.dart';
 
@@ -32,7 +33,7 @@ class KycBanner extends StatelessWidget {
             Icon(
               PhosphorIcons.shieldWarning(PhosphorIconsStyle.duotone),
               color: trustTheme.warning,
-              size: 24,
+              size: DeelmarktIconSize.md,
             ),
             const SizedBox(width: Spacing.s3),
             Expanded(
