+++ lib/features/auth/presentation/widgets/kyc_bottom_sheet.dart
@@ -5,6 +5,7 @@ import 'package:phosphor_flutter/phosphor_flutter.dart';
 
 import 'package:deelmarkt/core/design_system/animation.dart';
 import 'package:deelmarkt/core/design_system/colors.dart';
+import 'package:deelmarkt/core/design_system/icon_sizes.dart';
 import 'package:deelmarkt/core/design_system/radius.dart';
 import 'package:deelmarkt/core/design_system/spacing.dart';
 import 'package:deelmarkt/features/auth/presentation/viewmodels/kyc_prompt_viewmodel.dart';
@@ -166,7 +167,7 @@ class KycBottomSheet extends ConsumerWidget {
             child: Icon(
               PhosphorIcons.checkCircle(PhosphorIconsStyle.fill),
               color: DeelmarktColors.trustVerified,
-              size: 64,
+              size: DeelmarktIconSize.hero,
             ),
           ),
           const SizedBox(height: Spacing.s4),
