+++ lib/features/auth/presentation/widgets/kyc_trust_footer.dart
@@ -2,6 +2,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 
+import 'package:deelmarkt/core/design_system/icon_sizes.dart';
 import 'package:deelmarkt/core/design_system/spacing.dart';
 
 /// Small trust footer at the bottom of the KYC sheet.
@@ -15,7 +16,7 @@ class KycTrustFooter extends StatelessWidget {
       children: [
         Icon(
           PhosphorIcons.shieldCheck(PhosphorIconsStyle.fill),
-          size: 16,
+          size: DeelmarktIconSize.xs,
           color: Theme.of(context).colorScheme.onSurfaceVariant,
         ),
         const SizedBox(width: Spacing.s1),
