+++ lib/features/home/presentation/widgets/seller_stats_row.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 
 import 'package:deelmarkt/core/design_system/colors.dart';
+import 'package:deelmarkt/core/design_system/icon_sizes.dart';
 import 'package:deelmarkt/core/design_system/radius.dart';
 import 'package:deelmarkt/core/design_system/spacing.dart';
 import 'package:deelmarkt/core/utils/formatters.dart';
@@ -95,7 +96,7 @@ class _StatCard extends StatelessWidget {
           children: [
             Row(
               children: [
-                Icon(icon, size: 20, color: iconColor),
+                Icon(icon, size: DeelmarktIconSize.sm, color: iconColor),
                 if (showBadge) ...[
                   const SizedBox(width: Spacing.s1),
                   Container(
