+++ lib/features/messages/presentation/chat_thread_providers.dart
@@ -5,14 +5,6 @@ import 'package:deelmarkt/features/messages/domain/usecases/get_messages_usecase
 import 'package:deelmarkt/features/messages/domain/usecases/send_message_usecase.dart';
 import 'package:deelmarkt/features/messages/domain/usecases/update_offer_status_usecase.dart';
 
-/// Stub id for the current signed-in user.
-///
-/// TODO(auth): replace with `authStateProvider.currentUser.id` once the auth
-/// subsystem ships via the `[R]` backend tasks. This single constant is the
-/// source of truth for both the notifier (optimistic send sender) and the
-/// screen (self vs other bubble alignment).
-const String kCurrentUserIdStub = 'user-001';
-
 final getMessagesUseCaseProvider = Provider<GetMessagesUseCase>(
   (ref) => GetMessagesUseCase(ref.watch(messageRepositoryProvider)),
 );
