+++ lib/features/search/presentation/widgets/filter_bottom_sheet.dart
@@ -2,6 +2,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import 'package:deelmarkt/core/design_system/icon_sizes.dart';
 import 'package:deelmarkt/core/design_system/radius.dart';
 import 'package:deelmarkt/core/design_system/spacing.dart';
 import 'package:deelmarkt/features/search/presentation/search_providers.dart';
@@ -121,11 +122,11 @@ class _FilterSheetState extends ConsumerState<_FilterSheet> {
         ),
         const SizedBox(height: Spacing.s2),
         categoriesAsync.when(
-          loading: () => const SizedBox(height: 48),
+          loading: () => const SizedBox(height: Spacing.s12),
           error:
               (_, _) => TextButton.icon(
                 onPressed: () => ref.invalidate(topLevelCategoriesProvider),
-                icon: const Icon(Icons.refresh, size: 16),
+                icon: const Icon(Icons.refresh, size: DeelmarktIconSize.xs),
                 label: Text('action.retry'.tr()),
               ),
           data:
